Moved JwtAuthFilter to filters package
